requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
email-validator==2.0.1
pytest==7.3.1
python-dotenv==1.0.0
flask-dance[sqla]==7.0.0
